Add code for reading MHS and XMP files
This is pretty much the only part of the project that is in a fully
working state, as the conversion code is far from finished and relies on
BlockDesign stuff which has not been written yet.

These files are slightly older versions that I was able to restore after
accidentally deleting everything. I have done my best to redo any
changes that were lost, but I might have missed some. Still, everything
seems to be working.